Remove TODO Comment
I decided to keep the alternating white/grey even though it makes the
contact section white when it is grey in the mockup because of the
added skills section. I feel it is more important to keep the pattern
of alternating background colors to distinguish between the sections
and I am assuming that this is the spirit of the mockup anyway.