Added Large Screen Breakpoint
Breakpoint for 1024px and above added. I set the max width to 1000px
because it appears that the mockup had a similar max width. However, I
changed my page to allow the grey/white alternating background to
extend to the full width of the page while still limiting content to
1000px. I also fixed some layout issues with the 768px breakpoint and
changed the text color of the skill paragraph to meet another ‘exceeds’
goal on the rubric.